+++ HoneyDrunk.Notify/HoneyDrunk.Notify.Queue.InMemory/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to HoneyDrunk.Notify.Queue.InMemory will be documented in th
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [0.2.0] - 2026-05-05
+
+### Changed
+
+- Aligned package version to `0.2.0` for the ADR-0019 Notify intake boundary refactor.
+
 ## [0.1.0] - 2026-01-01
 
 ### Added